further refactors client structure to allow graphical user interface.
A ClientUserCommandHandler is introduced as interface to workload mapper and
client-server connection which is separated from the user interface. The client
starts the user interface and passes its user command handler. As soon as the interface
returns (because of an error or because the user terminates the program) the client
closes all server connections and exits.

* FIXED: Execution map is now only created once and cleared and filled by the workload
mapper.