Energy and performance charts added
Energy and performance charts are displayed in the gui using the xcharts java library.
For each execution run total energy and runtime are claculated from the returned results
and added as new data point to the displayed graphs.

TODO: evaluation and display of error cases. Right now they are displayed in the console.
TODO: evaluation of different possible mappings.