extends user commands with execution device specification.
The mapper determines an optimal execution device for the execution platform, i.e. GPU or CPU. This
device configuration is transmitted to the server.

Extends KMEANS benchmark execution on server side to consider the configured device. As a quick fix
the KMEANS code is exepcted to be present as configured for GPU in the "kmeans_gpu" folder and as
configured for CPU in the "kmeans_cpu" folder.