adds predictive mapper
The predictive mapper assigns a benchmark to an execution device (CPU,GPU) for
a specific server architecture. Assignments depend on the benchmark's workload and
the benchmarks openCL kernel.

The assignment decision is based on a model built for the server architecture which
was trained using openCL kernel features. It is able to predict the most energy efficient
and runtime efficient execution device for the corresponding architecture.

For demo purposes, the predictive mapper does not execute live predictions in this version.
Instead it selects the optimal benchmark to server device mapping from a list of predictions
made offline ahead of time.