properly handles swing gui update and background work for a single benchmark execution
Swing ui updates need to be executed in the event dispatch thread while all background work
needs to be executed in a separate thread. The SwingWorker provides a worker thread for
background work and is used for a single benchmark execution now.

adds a button for automatic mode but no background work yet.