adds alternative mapping series to energy and runtime charts.
Depending on the server configurations, a graph series is added for each possible execution device
combination. For a two servers that would be CPU0_CPU1, GPU0_CPU1, CPU0_GPU1, GPU0_GPU1. It is assumed
for now that all servers have the same architecture and run the same workload and benchmarks. The additional
graphs are displayed to show how the mapper compares with all other possible mappings.

only average energy and runtime values of non interactive execution statistics are saved