adds support for different server architectures
The architecture id reported by the server upon connection with client is used to
find task to execution device mappings from model precalculations and to find
execution statistics for the graphical display of alternative mappings. An execution device
prediction model needs to be trained for each architecture description to consider
architecture differences.